backend/photomi/pyproject.toml
[project]
name = "photomi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.18.0",
    "fastapi[standard]>=0.68.0",
    "pillow>=8.3.0",
    "python-dotenv>=0.19.0",
    "python-multipart>=0.0.5",
    "sqlalchemy>=1.4.0",
    "uvicorn>=0.15.0",
]
